Extract agreements helpers into submodule
Convert backend/routes/agreements.py to a package. The 10 module-level
helpers move to backend/routes/agreements/helpers.py: the public-eligible
SQL expression, both query-building helpers
(_build_agreement_search_match_query, _apply_agreement_metadata_filters),
the _tax_clause_rows aggregator, snippet stripping, industry-label
normalization, the cacheable JSON response wrapper, and small utilities.

__init__.py drops from 2408 to 1945 LOC and re-exports the four helpers
that external modules (mcp/tools, tax_clauses_service, tests) already
import from this module so the public surface is unchanged.

Updates two test path lists (runtime typing guards, r2 allowlist) for
the new file layout.
