+++ backend/tests/test_runtime_typing_guards.py
@@ -9,7 +9,8 @@ _RUNTIME_FILES = [
     "backend/routes/auth/__init__.py",
     "backend/routes/auth/cookies.py",
     "backend/routes/auth/zitadel_config.py",
-    "backend/routes/agreements.py",
+    "backend/routes/agreements/__init__.py",
+    "backend/routes/agreements/helpers.py",
     "backend/routes/reference_data.py",
     "backend/routes/sections.py",
     "backend/services/sections_service.py",
